rover-backend/requirements.txt
wheel==0.36.2
Django==3.1.6
django-extensions==3.1.0
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-rest-auth[with_social]==0.9.5
psycopg2-binary==2.8.6
